requirements.txt
fastapi==0.129.0
uvicorn==0.40.0
sqlalchemy==2.0.46
asyncpg==0.31.0
greenlet==3.3.1
alembic==1.18.4
python-jose==3.5.0
bcrypt==5.0.0
python-socketio==5.16.1
pydantic==2.12.5
email-validator==2.3.0
pydantic-settings==2.12.0
python-dotenv==1.2.1
beautifulsoup4==4.14.3
requests==2.32.5
